+++ tapo/src/api.rs
@@ -1,7 +1,6 @@
 mod api_client;
 mod child_devices;
 mod color_light_handler;
-mod color_light_strip_handler;
 mod generic_device_handler;
 mod handler_ext;
 mod hub_handler;
@@ -10,11 +9,12 @@ mod plug_energy_monitoring_handler;
 mod plug_handler;
 mod power_strip_handler;
 mod protocol;
+mod rgb_light_strip_handler;
+mod rgbic_light_strip_handler;
 
 pub use api_client::*;
 pub use child_devices::*;
 pub use color_light_handler::*;
-pub use color_light_strip_handler::*;
 pub use generic_device_handler::*;
 pub use handler_ext::*;
 pub use hub_handler::*;
@@ -22,3 +22,5 @@ pub use light_handler::*;
 pub use plug_energy_monitoring_handler::*;
 pub use plug_handler::*;
 pub use power_strip_handler::*;
+pub use rgb_light_strip_handler::*;
+pub use rgbic_light_strip_handler::*;