+++ tapo/src/responses/device_info_result.rs
@@ -1,17 +1,19 @@
 mod color_light;
-mod color_light_strip;
 mod default_state;
 mod generic;
 mod hub;
 mod light;
 mod plug;
 mod power_strip;
+mod rgb_light_strip;
+mod rgbic_light_strip;
 
 pub use color_light::*;
-pub use color_light_strip::*;
 pub use default_state::*;
 pub use generic::*;
 pub use hub::*;
 pub use light::*;
 pub use plug::*;
 pub use power_strip::*;
+pub use rgb_light_strip::*;
+pub use rgbic_light_strip::*;