+++ tapo/src/responses/device_info_result/rgb_light_strip.rs
@@ -1,13 +1,12 @@
 use serde::{Deserialize, Serialize};
 
 use crate::error::Error;
-use crate::requests::LightingEffect;
 use crate::responses::{decode_value, DecodableResultExt, DefaultStateType, TapoResponseExt};
 
-/// Device info of Tapo L920 and L930. Superset of [`crate::responses::DeviceInfoGenericResult`].
+/// Device info of Tapo L900. Superset of [`crate::responses::DeviceInfoGenericResult`].
 #[derive(Debug, Clone, Serialize, Deserialize)]
 #[allow(missing_docs)]
-pub struct DeviceInfoColorLightStripResult {
+pub struct DeviceInfoRgbLightStripResult {
     //
     // Inherited from DeviceInfoGenericResult
     //
@@ -44,12 +43,12 @@ pub struct DeviceInfoColorLightStripResult {
     pub color_temp: u16,
     pub color_temp_range: [u16; 2],
     /// The default state of a device to be used when internet connectivity is lost after a power cut.
-    pub default_states: DefaultColorLightStripState,
+    pub default_states: DefaultRgbLightStripState,
 }
 
-impl TapoResponseExt for DeviceInfoColorLightStripResult {}
+impl TapoResponseExt for DeviceInfoRgbLightStripResult {}
 
-impl DecodableResultExt for DeviceInfoColorLightStripResult {
+impl DecodableResultExt for DeviceInfoRgbLightStripResult {
     fn decode(mut self) -> Result<Self, Error> {
         self.ssid = decode_value(&self.ssid)?;
         self.nickname = decode_value(&self.nickname)?;
@@ -61,18 +60,17 @@ impl DecodableResultExt for DeviceInfoColorLightStripResult {
 /// Color Light Strip Default State.
 #[derive(Debug, Clone, Serialize, Deserialize)]
 #[allow(missing_docs)]
-pub struct DefaultColorLightStripState {
+pub struct DefaultRgbLightStripState {
     pub r#type: DefaultStateType,
-    pub state: ColorLightStripState,
+    pub state: RgbLightStripState,
 }
 
 /// Color Light Strip State.
 #[derive(Debug, Clone, Serialize, Deserialize)]
 #[allow(missing_docs)]
-pub struct ColorLightStripState {
+pub struct RgbLightStripState {
     pub brightness: Option<u8>,
     pub hue: Option<u16>,
     pub saturation: Option<u16>,
     pub color_temp: Option<u16>,
-    pub lighting_effect: Option<LightingEffect>,
 }